Created CoreDataStorageManager, remove core data from AppDelegate and VCs Added LaunchScreen
